pyproject.toml
[tool.poetry]
name = "my_commercial_project_1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pytelegrambotapi = "^4.14.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.lint.dependencies]
black = "^23.11.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
